feat(sql): Refactor test
